music/producer/requirements.txt
kafka-python
pandas
requests
